refactor: move utils to api
