+++ frontend/svelte/src/lib/services/dev.services.ts
@@ -1,10 +1,10 @@
 import type { BlockHeight, TransferError } from "@dfinity/nns";
 import { InsufficientFunds } from "@dfinity/nns";
 import { get } from "svelte/store";
+import { acquireICPTs } from "../api/dev.api";
 import { E8S_PER_ICP } from "../constants/icp.constants";
 import { AccountsStore, accountsStore } from "../stores/accounts.store";
 import { AuthStore, authStore } from "../stores/auth.store";
-import { acquireICPTs } from "../utils/dev.utils";
 import { syncAccounts } from "./accounts.services";
 
 export const getICPs = async (icps: number) => {
